+++ src/cs-utils/utils.cpp
@@ -131,7 +131,7 @@ std::string exec(std::string const& cmd) {
   if (!pipe) {
     throw std::runtime_error("popen() failed!");
   }
-  while (fgets(buffer.data(), buffer.size(), pipe.get()) != nullptr) {
+  while (fgets(buffer.data(), static_cast<int>(buffer.size()), pipe.get()) != nullptr) {
     result += buffer.data();
   }
   return result;
